+++ CHANGELOG.md
@@ -2,6 +2,7 @@ QuantumAnnealing.jl Change Log
 ==============================
 
 ### Staged
+- Add tools for working with classical Ising models (#8)
 - Add data processing tools (#6)
 
 ### v0.0.1